separate subject metadata from behavior metadata
