+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/full_genotype.py
