+++ allensdk/brain_observatory/behavior/data_objects/metadata/subject_metadata/sex.py
